refactor(shipping): CHECKOUT-3791 Move shipping strategy options interface
