+++ src/shipping/strategies/amazon/amazon-pay-shipping-strategy.spec.ts
@@ -3,25 +3,25 @@ import { createRequestSender } from '@bigcommerce/request-sender';
 import { createScriptLoader } from '@bigcommerce/script-loader';
 import { of } from 'rxjs';
 
-import { ConsignmentRequestSender } from '..';
-import { createCheckoutStore, CheckoutRequestSender, CheckoutStore, CheckoutStoreState } from '../../checkout';
-import { getCheckoutStoreState } from '../../checkout/checkouts.mock';
-import { InvalidArgumentError, MissingDataError } from '../../common/error/errors';
-import { PaymentMethodActionCreator, PaymentMethodActionType, PaymentMethodRequestSender } from '../../payment';
-import { getAmazonPay } from '../../payment/payment-methods.mock';
+import { ConsignmentRequestSender } from '../..';
+import { createCheckoutStore, CheckoutRequestSender, CheckoutStore, CheckoutStoreState } from '../../../checkout';
+import { getCheckoutStoreState } from '../../../checkout/checkouts.mock';
+import { InvalidArgumentError, MissingDataError } from '../../../common/error/errors';
+import { PaymentMethodActionCreator, PaymentMethodActionType, PaymentMethodRequestSender } from '../../../payment';
+import { getAmazonPay } from '../../../payment/payment-methods.mock';
 import {
     AmazonPayAddressBook,
     AmazonPayAddressBookOptions,
     AmazonPayOrderReference,
     AmazonPayScriptLoader,
     AmazonPayWindow,
-} from '../../payment/strategies/amazon-pay';
-import { RemoteCheckoutActionCreator, RemoteCheckoutActionType, RemoteCheckoutRequestSender } from '../../remote-checkout';
-import ConsignmentActionCreator from '../consignment-action-creator';
-import { ConsignmentActionType } from '../consignment-actions';
-import { getFlatRateOption } from '../internal-shipping-options.mock';
-import { getShippingAddress } from '../shipping-addresses.mock';
-import { ShippingStrategyActionType } from '../shipping-strategy-actions';
+} from '../../../payment/strategies/amazon-pay';
+import { RemoteCheckoutActionCreator, RemoteCheckoutActionType, RemoteCheckoutRequestSender } from '../../../remote-checkout';
+import ConsignmentActionCreator from '../../consignment-action-creator';
+import { ConsignmentActionType } from '../../consignment-actions';
+import { getFlatRateOption } from '../../internal-shipping-options.mock';
+import { getShippingAddress } from '../../shipping-addresses.mock';
+import { ShippingStrategyActionType } from '../../shipping-strategy-actions';
 
 import AmazonPayShippingStrategy from './amazon-pay-shipping-strategy';
