+++ src/shipping/strategies/amazon/index.ts
@@ -1,0 +1,2 @@
+export { default as AmazonPayShippingStrategy } from './amazon-pay-shipping-strategy';
+export { default as AmazonPayShippingInitializeOptions } from './amazon-pay-shipping-initialize-options';
