+++ src/shipping/strategies/default/default-shipping-strategy.ts
@@ -1,9 +1,9 @@
-import { AddressRequestBody } from '../../address';
-import { CheckoutStore, InternalCheckoutSelectors } from '../../checkout';
-import ConsignmentActionCreator from '../consignment-action-creator';
-import { ShippingRequestOptions } from '../shipping-request-options';
+import { AddressRequestBody } from '../../../address';
+import { CheckoutStore, InternalCheckoutSelectors } from '../../../checkout';
+import ConsignmentActionCreator from '../../consignment-action-creator';
+import { ShippingRequestOptions } from '../../shipping-request-options';
 
-import ShippingStrategy from './shipping-strategy';
+import ShippingStrategy from '../shipping-strategy';
 
 export default class DefaultShippingStrategy implements ShippingStrategy {
     constructor(
