+++ src/shipping/strategies/default/index.ts
@@ -1,0 +1,1 @@
+export { default as DefaultShippingStrategy } from './default-shipping-strategy';
