+++ src/shipping/strategies/index.ts
@@ -1,3 +1,1 @@
-export { default as AmazonPayShippingStrategy, AmazonPayShippingInitializeOptions } from './amazon-pay-shipping-strategy';
-export { default as DefaultShippingStrategy } from './default-shipping-strategy';
 export { default as ShippingStrategy } from './shipping-strategy';
